Change time module to trips_info
- change name of time module to trips_info
-  update introduction of trips_info
- move get_time, get_duration, get_position functions from trips_input to trips_info
- test all functions: worked
